3.27

+++ app/Http/Controllers/Api/TweetController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateTweetRequest;
 use Illuminate\Http\Request;
 use App\Models\Tweet;
 use App\Services\TweetService;
+use Illuminate\Support\Facades\Gate;
 
 class TweetController extends Controller
 {
@@ -25,7 +26,8 @@ class TweetController extends Controller
    */
   public function index()
   {
-    // 🔽 編集
+    Gate::authorize('viewAny', Tweet::class);
+
     $tweets = $this->tweetService->allTweets();
     return response()->json($tweets);
   }
@@ -35,8 +37,9 @@ class TweetController extends Controller
    */
   public function store(StoreTweetRequest $request)
   {
-    $tweet = $this->tweetService->createTweet($request);
+    Gate::authorize('create', Tweet::class);
 
+    $tweet = $this->tweetService->createTweet($request);
     return response()->json($tweet, 201);
   }
 
@@ -45,6 +48,8 @@ class TweetController extends Controller
    */
   public function show(Tweet $tweet)
   {
+    Gate::authorize('view', $tweet);
+    
     return response()->json($tweet);
   }
 
@@ -53,8 +58,8 @@ class TweetController extends Controller
    */
   public function update(UpdateTweetRequest $request, Tweet $tweet)
   {
+    Gate::authorize('update', $tweet);
     $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
-
     return response()->json($updatedTweet);
   }
   /**
@@ -62,7 +67,8 @@ class TweetController extends Controller
    */
   public function destroy(Tweet $tweet)
   {
-    // 🔽 編集
+    Gate::authorize('delete', $tweet);
+    
     $this->tweetService->deleteTweet($tweet);
     return response()->json(['message' => 'Tweet deleted successfully']);
   }
