+++ app/Http/Controllers/TweetController.php
@@ -7,14 +7,13 @@ use App\Http\Requests\UpdateTweetRequest;
 use App\Models\Tweet;
 use Illuminate\Http\Request;
 use App\Services\TweetService;
+use Illuminate\Support\Facades\Gate;
 
 
 class TweetController extends Controller
 {
-  // 🔽 追加
   protected $tweetService;
 
-  // 🔽 追加
   public function __construct(TweetService $tweetService)
   {
     $this->tweetService = $tweetService;
@@ -25,7 +24,8 @@ class TweetController extends Controller
    */
   public function index()
   {
-    // 🔽 編集
+    Gate::authorize('viewAny', Tweet::class);
+
     $tweets = $this->tweetService->allTweets();
     return view('tweets.index', compact('tweets'));
   }
@@ -35,6 +35,7 @@ class TweetController extends Controller
    */
   public function create()
   {
+    Gate::authorize('create', Tweet::class);
     return view('tweets.create');
   }
 
@@ -43,8 +44,9 @@ class TweetController extends Controller
    */
   public function store(StoreTweetRequest $request)
   {
-    $tweet = $this->tweetService->createTweet($request);
+    Gate::authorize('create', Tweet::class);
 
+    $tweet = $this->tweetService->createTweet($request);
     return redirect()->route('tweets.index');
   }
 
@@ -53,6 +55,7 @@ class TweetController extends Controller
    */
   public function show(Tweet $tweet)
   {
+    Gate::authorize('view', $tweet);
     return view('tweets.show', compact('tweet'));
   }
 
@@ -61,6 +64,8 @@ class TweetController extends Controller
    */
   public function edit(Tweet $tweet)
   {
+    Gate::authorize('update', $tweet);
+
     return view('tweets.edit', compact('tweet'));
   }
 
@@ -69,8 +74,9 @@ class TweetController extends Controller
    */
   public function update(UpdateTweetRequest $request, Tweet $tweet)
   {
-    $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
+    Gate::authorize('update', $tweet);
 
+    $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
     return redirect()->route('tweets.show', $tweet);
   }
 
@@ -79,9 +85,9 @@ class TweetController extends Controller
    */
   public function destroy(Tweet $tweet)
   {
-    // 🔽 編集
-    $this->tweetService->deleteTweet($tweet);
+    Gate::authorize('delete', $tweet);
 
+    $this->tweetService->deleteTweet($tweet);
     return redirect()->route('tweets.index');
   }
 }
